replace main by sdl main
also only keep fb canvas and reorganize the code
