+++ src/lib.rs
@@ -4,3 +4,4 @@ pub mod geometry;
 pub mod math;
 pub mod render;
 pub mod scene;
+pub mod sdl;
